Build kubectl-directcsi as a single binary along with direct-csi
- Use `go build -tags "plugin" -o kubectl-directcsi` to build the plugin
- Use `go build` to build direct-csi
