+++ hack/build-without-docker.sh
@@ -26,3 +26,4 @@ export CGO_ENABLED=0
 
 "${SCRIPT_ROOT}/update-codegen.sh"
 go build -tags "osusergo netgo static_build" -ldflags="-X ${REPOSITORY}/cmd.Version=${CSI_VERSION} -extldflags=-static"
+go build -tags "plugin" -ldflags="-X ${REPOSITORY}/cmd.Version=${CSI_VERSION} -extldflags=-static" -o kubectl-directcsi
